requirements.txt
discord.py==1.3.4
pytz==2021.1
requests==2.24.0
kubernetes==23.3.0
